Extract code from ExpenditureApp.xaml window into user controls
This will allow flexibility in the extraction, allowing different views to be switched in and out based on the extraction mode. This will be more elegant that hiding and showing a host of controls.
