+++ ExpenditureAppWPFView/InputUserControl.xaml
@@ -1,12 +1,11 @@
-﻿<Page x:Class="ExpenditureAppWPF.InputPage"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:ExpenditureAppWPF"
-      mc:Ignorable="d" 
-        Title="MainWindow" Height="400" Width="885">
-
+﻿<UserControl x:Class="ExpenditureAppWPF.InputUserControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:ExpenditureAppWPF"
+             mc:Ignorable="d" 
+             d:DesignHeight="400" d:DesignWidth="885">
     <Grid>
         <Grid.RowDefinitions>
             <RowDefinition Height="Auto" />
@@ -175,4 +174,4 @@
             </Grid>
         </Grid>
     </Grid>
-</Page>
+</UserControl>
