+++ ExpenditureAppWPFView/OutputUserControl1.xaml
@@ -1,0 +1,67 @@
+﻿<UserControl x:Class="ExpenditureAppWPF.OutputUserControl1"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:ExpenditureAppWPF"
+             mc:Ignorable="d" 
+             d:DesignHeight="400" d:DesignWidth="885">
+    <Grid>
+        <Grid Margin="0,20,0,0">
+            <Grid.RowDefinitions>
+                <RowDefinition Height="Auto" />
+                <RowDefinition Height="Auto" />
+                <RowDefinition Height="Auto" />
+                <RowDefinition Height="*" />
+            </Grid.RowDefinitions>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="Auto" />
+                <ColumnDefinition Width="Auto" />
+                <ColumnDefinition Width="*" />
+            </Grid.ColumnDefinitions>
+            <StackPanel>
+                <StackPanel DockPanel.Dock="Top" Orientation="Horizontal">
+                    <Label VerticalAlignment="Bottom">Mode</Label>
+                    <RadioButton VerticalAlignment="Center" Margin="5">By Tag</RadioButton>
+                    <RadioButton VerticalAlignment="Center" Margin="5">Total</RadioButton>
+                </StackPanel>
+                <ComboBox Margin="5"></ComboBox>
+            </StackPanel>
+            <StackPanel Grid.Column="1" Orientation="Horizontal">
+                <StackPanel Margin="0,0,5,0">
+                    <Label>Start Date</Label>
+                    <StackPanel Orientation="Horizontal">
+                        <TextBox Margin="5" MinWidth="20"></TextBox>
+                        <TextBox Margin="5" MinWidth="20"></TextBox>
+                        <TextBox Margin="5" MinWidth="20"></TextBox>
+                    </StackPanel>
+                </StackPanel>
+                <StackPanel Margin="5,0,0,0">
+                    <Label>End Date</Label>
+                    <StackPanel Orientation="Horizontal">
+                        <TextBox Margin="5" MinWidth="20"></TextBox>
+                        <TextBox Margin="5" MinWidth="20"></TextBox>
+                        <TextBox Margin="5" MinWidth="20"></TextBox>
+                        <Button MinWidth="40" Margin="10,5,0,5">Go</Button>
+                    </StackPanel>
+                </StackPanel>
+            </StackPanel>
+            <ComboBox Grid.Column="1" Grid.Row="1" Margin="5,5,55,5"></ComboBox>
+            <StackPanel Grid.Row="2">
+                <Label DockPanel.Dock="Top">Dominant Tag</Label>
+                <ComboBox DockPanel.Dock="Bottom" Margin="5"></ComboBox>
+            </StackPanel>
+            <StackPanel Grid.Row="2" Grid.Column="1">
+                <Label DockPanel.Dock="Top">Associated Tag</Label>
+                <ComboBox DockPanel.Dock="Bottom" Margin="5,5,80,5"></ComboBox>
+            </StackPanel>
+            <StackPanel Grid.Row="2" Grid.Column="2">
+                <Label DockPanel.Dock="Top" Margin="-65,0,380,0">Person</Label>
+                <ComboBox DockPanel.Dock="Bottom" Margin="-65,5,380,5"></ComboBox>
+            </StackPanel>
+            <Button Grid.Column="2" Grid.Row="3" HorizontalAlignment="Right" VerticalAlignment="Bottom" Margin="5">
+                Reset Fields
+            </Button>
+        </Grid>
+    </Grid>
+</UserControl>
